[poincare] Integer: revert uint8_t--> size_t wrongly changed

+++ poincare/include/poincare/integer.h
@@ -35,7 +35,7 @@ public:
   Integer(native_int_t i = 0);
   Integer(double_native_int_t i);
   Integer(native_uint_t * digits, uint16_t numberOfDigits, bool negative, bool enableOverflow = false);
-  Integer(const char * digits, uint8_t length, bool negative);
+  Integer(const char * digits, size_t length, bool negative);
   Integer(const char * digits) : Integer(digits, strlen(digits), false) {}
   static Integer Overflow(bool negative) { return Integer((native_uint_t *)nullptr, k_maxNumberOfDigits+1, negative); }
   ~Integer();
